+++ src/Exception/TokenNotFoundException.php
@@ -4,5 +4,4 @@ namespace ItkDev\OpenIdConnectBundle\Exception;
 
 class TokenNotFoundException extends ItkOpenIdConnectBundleException
 {
-
 }