+++ src/Exception/UserDoesNotExistException.php
@@ -4,5 +4,4 @@ namespace ItkDev\OpenIdConnectBundle\Exception;
 
 class UserDoesNotExistException extends ItkOpenIdConnectBundleException
 {
-
 }