+++ src/Exception/UsernameDoesNotExistException.php
@@ -4,5 +4,4 @@ namespace ItkDev\OpenIdConnectBundle\Exception;
 
 class UsernameDoesNotExistException extends ItkOpenIdConnectBundleException
 {
-
 }