+++ src/Security/CliLoginTokenAuthenticator.php
@@ -1,0 +1,76 @@
+<?php
+
+namespace ItkDev\OpenIdConnectBundle\Security;
+
+use ItkDev\OpenIdConnectBundle\Exception\CacheException;
+use ItkDev\OpenIdConnectBundle\Exception\TokenNotFoundException;
+use ItkDev\OpenIdConnectBundle\Exception\UsernameDoesNotExistException;
+use ItkDev\OpenIdConnectBundle\Util\CliLoginHelper;
+use Symfony\Component\HttpFoundation\RedirectResponse;
+use Symfony\Component\HttpFoundation\Request;
+use Symfony\Component\HttpFoundation\Response;
+use Symfony\Component\Routing\Generator\UrlGeneratorInterface;
+use Symfony\Component\Security\Core\Authentication\Token\TokenInterface;
+use Symfony\Component\Security\Core\Exception\AuthenticationException;
+use Symfony\Component\Security\Core\Exception\CustomUserMessageAuthenticationException;
+use Symfony\Component\Security\Core\User\UserProviderInterface;
+use Symfony\Component\Security\Http\Authenticator\AbstractAuthenticator;
+use Symfony\Component\Security\Http\Authenticator\Passport\Badge\UserBadge;
+use Symfony\Component\Security\Http\Authenticator\Passport\SelfValidatingPassport;
+
+/**
+ * Authenticator class for CLI login.
+ */
+class CliLoginTokenAuthenticator extends AbstractAuthenticator
+{
+    private CliLoginHelper $cliLoginHelper;
+    private UserProviderInterface $userProvider;
+    private UrlGeneratorInterface $router;
+
+    private string $cliLoginRedirectRoute;
+
+    public function __construct(CliLoginHelper $cliLoginHelper, UserProviderInterface $userProvider, string $cliLoginRedirectRoute, UrlGeneratorInterface $router)
+    {
+        $this->cliLoginHelper = $cliLoginHelper;
+        $this->userProvider = $userProvider;
+        $this->cliLoginRedirectRoute = $cliLoginRedirectRoute;
+        $this->router = $router;
+    }
+
+    public function supports(Request $request): ?bool
+    {
+        return $request->query->has('loginToken');
+    }
+
+    public function authenticate(Request $request)
+    {
+        $token = (string) $request->query->get('loginToken');
+        if (empty($token)) {
+            // The token header was empty, authentication fails with HTTP Status
+            // Code 401 "Unauthorized"
+            throw new CustomUserMessageAuthenticationException('No login token provided');
+        }
+
+        try {
+            $username = $this->cliLoginHelper->getUsername($token);
+        } catch (CacheException | TokenNotFoundException $e) {
+            throw new CustomUserMessageAuthenticationException('Cannot get username');
+        }
+
+        if (null === $username) {
+            throw new UsernameDoesNotExistException('null is not a valid username.');
+        }
+
+        return new SelfValidatingPassport(new UserBadge($username));
+    }
+
+    public function onAuthenticationSuccess(Request $request, TokenInterface $token, string $firewallName): ?Response
+    {
+        return new RedirectResponse($this->router->generate($this->cliLoginRedirectRoute));
+    }
+
+    public function onAuthenticationFailure(Request $request, AuthenticationException $exception): ?Response
+    {
+        throw new AuthenticationException('Error occurred validating login token');
+    }
+}