+++ src/Security/OpenIdLoginAuthenticator.php
@@ -4,52 +4,48 @@ namespace ItkDev\OpenIdConnectBundle\Security;
 
 use ItkDev\OpenIdConnect\Exception\ItkOpenIdConnectException;
 use ItkDev\OpenIdConnect\Exception\ValidationException;
-use ItkDev\OpenIdConnect\Security\OpenIdConfigurationProvider;
 use ItkDev\OpenIdConnectBundle\Exception\InvalidProviderException;
 use Symfony\Component\HttpFoundation\Request;
+use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\HttpFoundation\Session\SessionInterface;
-use Symfony\Component\Security\Core\Authentication\Token\TokenInterface;
 use Symfony\Component\Security\Core\Exception\AuthenticationException;
-use Symfony\Component\Security\Core\User\UserInterface;
-use Symfony\Component\Security\Core\User\UserProviderInterface;
-use Symfony\Component\Security\Guard\AbstractGuardAuthenticator;
+use Symfony\Component\Security\Http\Authenticator\AbstractAuthenticator;
+use Symfony\Component\Security\Http\EntryPoint\AuthenticationEntryPointInterface;
 
 /**
  * Authenticator for OpenId Connect login.
  */
-abstract class OpenIdLoginAuthenticator extends AbstractGuardAuthenticator
+abstract class OpenIdLoginAuthenticator extends AbstractAuthenticator implements AuthenticationEntryPointInterface
 {
-    /**
-     * @var SessionInterface
-     */
-    private $session;
+    private SessionInterface $session;
+    private OpenIdConfigurationProviderManager $providerManager;
 
-    /**
-     * @var OpenIdConfigurationProviderManager
-     */
-    private $providerManager;
-
-    private $leeway;
-
-    public function __construct(OpenIdConfigurationProviderManager $providerManager, SessionInterface $session, int $leeway = 0)
-    {
+    public function __construct(
+        OpenIdConfigurationProviderManager $providerManager,
+        SessionInterface $session
+    ) {
         $this->providerManager = $providerManager;
         $this->session = $session;
-        $this->leeway = $leeway;
     }
 
-    public function supports(Request $request)
+    public function supports(Request $request): ?bool
     {
         // Check if request has state and id_token
         return $request->query->has('state') && $request->query->has('id_token');
     }
 
     /**
+     * @param Request $request
+     *
+     * @return array|string[]
+     *
+     * @throws ItkOpenIdConnectException
      * @throws ValidationException
+     * @throws InvalidProviderException
      */
-    public function getCredentials(Request $request)
+    protected function validateClaims(Request $request)
     {
-        $providerKey = (string)$this->session->remove('oauth2provider');
+        $providerKey = (string) $this->session->remove('oauth2provider');
         $provider = $this->providerManager->getProvider($providerKey);
 
         // Make sure state and oauth2state are the same
@@ -71,7 +67,7 @@ abstract class OpenIdLoginAuthenticator extends AbstractGuardAuthenticator
                 throw new ValidationException('Id token not type string');
             }
 
-            $claims = $provider->validateIdToken($idToken, $this->session->get('oauth2nonce'), $this->leeway);
+            $claims = $provider->validateIdToken($idToken, $this->session->get('oauth2nonce'));
             // Authentication successful
         } catch (ItkOpenIdConnectException $exception) {
             // Handle failed authentication
@@ -83,24 +79,8 @@ abstract class OpenIdLoginAuthenticator extends AbstractGuardAuthenticator
         return (array) $claims + ['open_id_connect_provider' => $providerKey];
     }
 
-    abstract public function getUser($credentials, UserProviderInterface $userProvider);
-
-    public function checkCredentials($credentials, UserInterface $user)
-    {
-        return true;
-    }
-
-    public function onAuthenticationFailure(Request $request, AuthenticationException $exception)
+    public function onAuthenticationFailure(Request $request, AuthenticationException $exception): ?Response
     {
         throw new AuthenticationException('Error occurred validating openid login');
-    }
-
-    abstract public function onAuthenticationSuccess(Request $request, TokenInterface $token, string $providerKey);
-
-    abstract public function start(Request $request, AuthenticationException $authException = null);
-
-    public function supportsRememberMe()
-    {
-        return false;
     }
 }