+++ CHANGELOG.md
@@ -5,8 +5,30 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html)
 
-### Added
+## [Unreleased]
 
+## [2.0.0] - 2021-12-08
+### Added
+- Migrated to Symfony's new (5.1+) security system
+### Changed
+- Require Symfony 5.4
+- Moved `leeway` config to provider config
+- ITK OpenID Connect: Upgraded from
+  `itk-dev/openid-connect` 2.1.0 to 3.0.0
+### Removed
+- Remove support for PHP 7.3
+
+## [1.1.0] - 2021-12-08
+### Added
+- Support for multiple open id connect configuration providers
+
+## [1.0.1] - 2021-09-20
+### Fixed
+- Updated README 
+- Avoided duplicate cache configuration
+
+## [1.0.0] - 2021-09-16
+### Added
 - README
 - LICENSE
 - OpenID Connect Bundle: Added bundle files, a login controller and an abstract authenticator.
@@ -17,6 +39,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Test Suite
 - Psalm setup for static analysis
 - Code formatting
-- OpenID Connect Bundle: Upgraded from
+- ITK OpenID Connect: Upgraded from
   `itk-dev/openid-connect` 1.0.0 to 2.1.0
 - OpenId Connect Bundle: Added CLI login feature.